Split out atoms, components, and themes packages

fix: adjust @polaris/atoms import
