+++ packages/atoms/src/index.ts
@@ -1,0 +1,2 @@
+export {atoms} from './atoms.css';
+export type {Atoms} from './atoms.css';
