+++ packages/components/src/Text/index.ts
