--- packages/components/src/atoms/index.ts
@@ -1,1 +1,0 @@
-export * from './atoms.css';
