--- packages/components/src/components/index.ts
@@ -1,10 +1,0 @@
-// Components
-export * from './Box';
-export * from './Center';
-export * from './Container';
-export * from './Flex';
-export * from './Grid';
-export * from './Inline';
-export * from './Link';
-export * from './Stack';
-export * from './Text';
