+++ packages/themes/README.md
@@ -1,0 +1,1 @@
+# @polaris/components
